backend/pyproject.toml
[tool.poetry]
name = "ideapace-backend"
version = "0.1.0"
description = "Backend for IdeaSpace"
authors = ["Nina Nikitenko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
redis = "^5.2.0"
channels = "^4.2.0"
celery = "^5.4.0"
kafka-python = "^2.0.2"
elasticsearch-dsl = {version = "^8.16.0", extras = ["search"]}
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
django-celery-beat = "^2.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
